// src/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
  id: string;
  title: string;
  slug: string;
  layout: (
    | {
        heading: string;
        tagline: string;
        image: string | Media;
        id?: string | null;
        blockName?: string | null;
        blockType: 'hero';
      }
    | {
        heading: string;
        details: string;
        stats?:
          | {
              label: string;
              value: number;
              suffix: string;
              id?: string | null;
              blockName?: string | null;
              blockType: 'stat';
            }[]
          | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'stats';
      }
    | {
        details: string;
        id?: string | null;
        blockName?: string | null;
        blockType: 'how-we-work';
      }
    | {
        heading: string;
        subheading: string;
        audio: string | Media;
        id?: string | null;
        blockName?: string | null;
        blockType: 'what-we-do';
      }
    | {
        heading: string;
        cards?:
          | {
              header: string;
              subheader: string;
              description: string;
              showActionButton?: boolean | null;
              actionButtonLink?: string | null;
              actionButtonText?: string | null;
              id?: string | null;
              blockName?: string | null;
              blockType: 'card';
            }[]
          | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'services';
      }
    | {
        heading: string;
        details: string;
        contactDetails?:
          | {
              label: string;
              icon: string | Media;
              link: string;
              value: string;
              id?: string | null;
              blockName?: string | null;
              blockType: 'contact-detail';
            }[]
          | null;
        image: string | Media;
        id?: string | null;
        blockName?: string | null;
        blockType: 'contact';
      }
    | {
        heading: string;
        details: string;
        image: string | Media;
        id?: string | null;
        blockName?: string | null;
        blockType: 'collaboration';
      }
  )[];
  updatedAt: string;
  createdAt: string;
}